Restored original status
