+++ build.gradle
@@ -1,5 +1,4 @@
 apply plugin: 'java'
-apply plugin: 'eclipse'
 
 repositories {
 	mavenCentral()
