--- src/main/java/net/davidtanzer/babysteps/TimerFactory.java
@@ -1,26 +1,0 @@
-/*  Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/
-
-package net.davidtanzer.babysteps;
-
-import net.davidtanzer.babysteps.ui.TimerPresentationModel;
-
-public abstract class TimerFactory {
-	private static TimerFactory defaultTimerFactory = new DefaultTimerFactory();
-
-	public static TimerFactory get() {
-		return defaultTimerFactory;
-	}
-	
-	public abstract Timer createTimer(long secondsInCycle, TimerPresentationModel presentationModel);
-}
