--- src/main/java/net/davidtanzer/babysteps/ui/TimerSoundsPlayer.java
@@ -1,48 +1,0 @@
-/*  Licensed under the Apache License, Version 2.0 (the "License");
-    you may not use this file except in compliance with the License.
-    You may obtain a copy of the License at
-
-        http://www.apache.org/licenses/LICENSE-2.0
-
-    Unless required by applicable law or agreed to in writing, software
-    distributed under the License is distributed on an "AS IS" BASIS,
-    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-    See the License for the specific language governing permissions and
-    limitations under the License.
-*/
-
-package net.davidtanzer.babysteps.ui;
-
-import javax.sound.sampled.AudioInputStream;
-import javax.sound.sampled.AudioSystem;
-import javax.sound.sampled.Clip;
-
-import net.davidtanzer.babysteps.BabystepsTimer;
-
-public class TimerSoundsPlayer {
-
-	private synchronized void playSound(final String url) {
-		new Thread(new Runnable() {
-			@Override
-			public void run() {
-				try {
-					Clip clip = AudioSystem.getClip();
-					AudioInputStream inputStream = AudioSystem.getAudioInputStream(
-							BabystepsTimer.class.getResourceAsStream("/"+url));
-					clip.open(inputStream);
-					clip.start();
-				} catch (Exception e) {
-					System.err.println(e.getMessage());
-				}
-			}
-		}).start();
-	}
-
-	public void playTenSecondsWarningSound() {
-		playSound("2166__suburban-grilla__bowl-struck.wav");
-	}
-	
-	public void playTimeElapsedInfoSound() {
-		playSound("32304__acclivity__shipsbell.wav");
-	}
-}
